Fix license
The license file should be named LICENSE so that it can be automatically parsed by github. You should also not repeat the license facts in README.md, as there is a separate file for that (let's save the user's time). In short, I use the same license type, but make it more convenient.